# customer-order-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: 1.0.0
paths:
  /api/v1/customer-orders:
    post:
      summary: Creates an order.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/creteOrderRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: invalid input
        '500':
          description: Unhandled error by server

    get:
      summary: List orders
      operationId: listOrders
      requestBody:
        parameters:
          - name: limit
            in: query
            required: false
            schema:
              type: integer
              default: 20
              example: 20
            description: Number of records to return
          - name: offset
            in: query
            required: false
            schema:
              type: integer
              default: 0
              example: 0
            description: Starting position for pagination
          - name: category
            in: query
            required: false
            schema:
              type: string
              example: 'B2C'
            description: Filter orders by category
        responses:
          "200":
            description: Orders retrieved successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/OrderListResponse'
          "400":
            description: Invalid request parameters
          "500":
            description: Internal server error

  /api/v1/customer-orders/{id}:
    get:
      summary: Get order by id .
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'cf24d5ce-f7be-43e7-85fb-a70de8654dfc'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not found
        '500':
          description: Unhandled error by server

components:
  schemas:
    CreteOrderRequest:
      type: object
      description: contains order details fro creating order
      properties:
        id:
          type: string
          example: '40b044b9-01c8-4e5f-a484-8fbe3fe351de'
        category:
          type: string
          example: 'B2C'
        customer:
          $ref: '#/components/schemas/Customer'
        site:
          $ref: '#/components/schemas/Site'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        createdAt:
          type: string
          format: date-time
          example: '2026-05-13T10:15:30+02:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-05-17T10:15:30+02:00'

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          example: '40b044b9-01c8-4e5f-a484-8fbe3fe351de'
        category:
          type: string
          example: "B2C"
        customer:
          $ref: '#/components/schemas/Customer'
        site:
          $ref: '#/components/schemas/Site'
        state:
          type: string
          example: "DRAFT"
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        createdAt:
          type: string
          format: date-time
          example: "2026-05-13T10:15:30+02:00"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-17T10:15:30+02:00"

    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    Customer:
      type: object
      description: contains customer details
      properties:
        id:
          type: string
          example: 'CUST-100'

    Site:
      type: object
      description: contains site details
      properties:
        id:
          type: string
          example: 'SITE-502'

    OrderItem:
      type: object
      properties:
        productOfferingId:
          type: string
          example: 'PO-30001'

        quantity:
          type: integer
          example: 2

    PaymentMethod:
      type: object
      properties:
        type:
          type: string
          example: 'DIRECT_DEBIT'
        iban:
          type: string
          description: Required if payment type is DIRECT_DEBIT
          example: 'DE89370400440532013000'